Split apart entry-table.tsx
