+++ apps/astro-app/src/pages/data/[databaseName]/[tableName]/entries/[...slug].astro
@@ -6,7 +6,7 @@ import type { DatabaseInfo, DescriptorInfo, TableInfo } from "@/types/data";
 import config from "@universal_root/config.yml";
 import Layout from "@/layouts/Layout.astro";
 
-import EntryTable from "@/components/data/entries/entry-table";
+import { EntryTable } from "@/components/data/entries/entry-table-page";
 import { Toaster } from "@/components/ui/sonner";
 import LoginDialog from "@/components/auth/login-dialog";
